gpu-resource-manager/requirements.txt
kubernetes==28.1.0
pyyaml==6.0.1
requests==2.31.0
prometheus-client==0.17.1
flask==2.3.3
flask-cors==4.0.0
dataclasses-json==0.6.1
pytest>=6.0
pytest-cov>=2.0 